requirements.txt
aiogram==3.18.0
aiosqlite==0.21.0
python-dotenv==1.0.1
aiohttp==3.11.13